Another attempt at getting the packages right
Also fixed README and some @param/@return documentation.
Fixed the code smells.
