+++ src/main/java/com/marklogic/mule/connector/api/types/AuthenticationType.java
@@ -15,23 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.connector.internal.provider;
+package com.marklogic.mule.connector.api.types;
 
-import com.marklogic.client.DatabaseClient;
+public enum AuthenticationType {
 
-public enum ConnectionType {
-    DIRECT {
-        @Override
-        public DatabaseClient.ConnectionType getMarkLogicConnectionType() {
-            return DatabaseClient.ConnectionType.DIRECT;
-        }
-    },
-    GATEWAY {
-        @Override
-        public DatabaseClient.ConnectionType getMarkLogicConnectionType() {
-            return DatabaseClient.ConnectionType.GATEWAY;
-        }
-    };
+    DIGEST,
+    BASIC,
+    MARKLOGIC_CLOUD,
+    CERTIFICATE,
+    KERBEROS,
+    SAML;
 
-    public abstract DatabaseClient.ConnectionType getMarkLogicConnectionType();
 }
