+++ src/main/java/com/marklogic/mule/connector/api/types/ConnectionType.java
@@ -15,15 +15,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.connector.internal.provider;
+package com.marklogic.mule.connector.api.types;
 
-public enum AuthenticationType {
+import com.marklogic.client.DatabaseClient;
 
-    DIGEST,
-    BASIC,
-    MARKLOGIC_CLOUD,
-    CERTIFICATE,
-    KERBEROS,
-    SAML;
+public enum ConnectionType {
+    DIRECT {
+        @Override
+        public DatabaseClient.ConnectionType getMarkLogicConnectionType() {
+            return DatabaseClient.ConnectionType.DIRECT;
+        }
+    },
+    GATEWAY {
+        @Override
+        public DatabaseClient.ConnectionType getMarkLogicConnectionType() {
+            return DatabaseClient.ConnectionType.GATEWAY;
+        }
+    };
 
+    public abstract DatabaseClient.ConnectionType getMarkLogicConnectionType();
 }
