+++ src/main/java/com/marklogic/mule/connector/api/types/DocumentFormat.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.connector.internal.provider;
+package com.marklogic.mule.connector.api.types;
 
 import com.marklogic.client.io.Format;
