+++ src/main/java/com/marklogic/mule/connector/internal/extension/MarkLogicConnector.java
@@ -15,9 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.connector.api.extension;
+package com.marklogic.mule.connector.internal.extension;
 
-import com.marklogic.mule.connector.api.config.Configuration;
+import com.marklogic.mule.connector.internal.config.Configuration;
 import com.marklogic.mule.connector.internal.error.ErrorType;
 import org.mule.runtime.api.meta.Category;
 import org.mule.runtime.extension.api.annotation.Configurations;
