+++ src/main/java/com/marklogic/mule/connector/internal/operation/QueryParameters.java
@@ -24,8 +24,8 @@ import com.marklogic.client.io.StringHandle;
 import com.marklogic.client.query.QueryDefinition;
 import com.marklogic.client.query.QueryManager;
 import com.marklogic.mule.connector.internal.Utilities;
-import com.marklogic.mule.connector.internal.provider.QueryFormat;
-import com.marklogic.mule.connector.internal.provider.QueryType;
+import com.marklogic.mule.connector.api.types.QueryFormat;
+import com.marklogic.mule.connector.api.types.QueryType;
 import org.mule.runtime.extension.api.annotation.param.NullSafe;
 import org.mule.runtime.extension.api.annotation.param.Optional;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
