+++ src/test/java/com/marklogic/mule/extension/AbstractFlowTester.java
@@ -17,7 +17,7 @@
  */
 package com.marklogic.mule.extension;
 
-import com.marklogic.mule.connector.internal.provider.DocumentAttributes;
+import com.marklogic.mule.connector.api.types.DocumentAttributes;
 import org.junit.Before;
 import org.mule.functional.junit4.MuleArtifactFunctionalTestCase;
 import org.mule.runtime.api.message.Message;
