+++ src/test/java/com/marklogic/mule/extension/DocumentData.java
@@ -17,7 +17,7 @@
  */
 package com.marklogic.mule.extension;
 
-import com.marklogic.mule.connector.internal.provider.DocumentAttributes;
+import com.marklogic.mule.connector.api.types.DocumentAttributes;
 
 public class DocumentData {
