+++ src/test/java/com/marklogic/mule/extension/MetadataVerifier.java
@@ -17,7 +17,7 @@
  */
 package com.marklogic.mule.extension;
 
-import com.marklogic.mule.connector.internal.provider.DocumentAttributes;
+import com.marklogic.mule.connector.api.types.DocumentAttributes;
 import org.junit.Assert;
 
 import javax.xml.namespace.QName;
