+++ src/test/java/com/marklogic/mule/extension/SearchDocumentsWithMetadataTest.java
@@ -17,7 +17,7 @@
  */
 package com.marklogic.mule.extension;
 
-import com.marklogic.mule.connector.internal.provider.DocumentAttributes;
+import com.marklogic.mule.connector.api.types.DocumentAttributes;
 import org.junit.Test;
 
 import java.util.List;
