+++ src/test/java/com/marklogic/mule/extension/WriteDocumentTest.java
@@ -18,7 +18,7 @@
 package com.marklogic.mule.extension;
 
 
-import com.marklogic.mule.connector.internal.provider.DocumentAttributes;
+import com.marklogic.mule.connector.api.types.DocumentAttributes;
 import org.junit.Test;
 
 import java.util.HashSet;
